+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Enabled CI on multiple unity versions.
 - Changed projects settings to conform to Meta XR SDK recommendations.
+- BREAKING: Removed file format v2 distinguishing
 
 ## [0.1.1] - 2024.01.09